enemies shoot

+++ DOOM/bullets/imp_bullet.h
@@ -1,8 +1,7 @@
 #pragma once
 #include "bullet.h"
-#include "../animation/animation.h" 
 
 struct Imp_bullet : public Bullet
 {
-	inline Imp_bullet(int pos_x, int pos_y, int angle, int speed): Bullet(pos_x, pos_y, angle, speed) {};
+	inline Imp_bullet(float pos_x, float pos_y, float angle, float speed): Bullet(pos_x, pos_y, angle, speed) {};
 };
